Move all code into src/ folder
