+++ src/commands/docs.js
