+++ src/commands/keys.js
