+++ src/commands/logout.js
