+++ src/commands/signup.js
