+++ src/commands/update.js
