+++ src/commands/whoami.js
