+++ src/lib/package-json.js
