+++ src/utils/stub.js
