+++ test/commands/run.test.js
@@ -1,8 +1,8 @@
 const assert = require('assert');
 const nock = require('nock');
 
-const { BUILD_URL } = require('../../utils/utils');
-const { run } = require('../../commands/run');
+const { BUILD_URL } = require('../../src/utils/utils');
+const { run } = require('../../src/commands/run');
 
 describe('run command', () => {
   it('should throw if no service provided', () => {
